Added PSR-4 standard comply
